# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forecasting Competitive Football API",
    "summary": "Frozen-artifact inference for leakage-safe football forecasts",
    "description": "Versioned access to the finished project's Model 1 pre-match outcome probabilities, Model 2 signed goal margin, and Model 3 five-minute historical in-play snapshots. The service loads frozen artifacts once and never trains on startup. GitHub Pages hosts the static documentation only; this Python API must run separately.",
    "contact": {
      "name": "Project repository",
      "url": "https://github.com/Mana-Peiravian/forecasting-competitive-football"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Configured API server (local by default)"
    }
  ],
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Search the bounded public match catalogue",
        "operationId": "matches_api_v1_matches_get",
        "parameters": [
          {
            "name": "competition",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`La Liga` in this release.",
              "title": "Competition"
            },
            "description": "`La Liga` in this release."
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "2016/2017"
              ],
              "title": "Season"
            }
          },
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "train",
                    "validation",
                    "calibration",
                    "pilot_holdout",
                    "final_test"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Split"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get safe metadata for one match",
        "operationId": "match_api_v1_matches__match_id__get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Known StatsBomb match identifier",
              "title": "Match Id"
            },
            "description": "Known StatsBomb match identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict/prematch/outcome": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Run Model 1 pre-match outcome inference",
        "description": "Return calibrated P(Home), P(Draw), and P(Away) from frozen kickoff data.",
        "operationId": "predict_prematch_outcome_api_v1_predict_prematch_outcome_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrematchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrematchOutcomeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict/prematch/margin": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Run Model 2 signed goal-margin inference",
        "description": "Return expected clipped home-minus-away final goal margin.",
        "operationId": "predict_prematch_margin_api_v1_predict_prematch_margin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrematchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrematchMarginResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict/prematch": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Run the combined Model 1 and Model 2 forecast",
        "operationId": "predict_prematch_api_v1_predict_prematch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrematchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedPrematchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict/inplay": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Run Model 3 at a frozen five-minute snapshot",
        "description": "Use only the saved features whose events are at or before the snapshot boundary.",
        "operationId": "predict_inplay_api_v1_predict_inplay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InPlayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InPlayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}/timeline": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Replay all 19 Model 3 snapshots for one historical match",
        "operationId": "timeline_api_v1_matches__match_id__timeline_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Known StatsBomb match identifier",
              "title": "Match Id"
            },
            "description": "Known StatsBomb match identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CombinedPrematchResponse": {
        "properties": {
          "match_id": {
            "type": "integer",
            "title": "Match Id"
          },
          "match_date": {
            "type": "string",
            "format": "date",
            "title": "Match Date"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "split": {
            "type": "string",
            "enum": [
              "train",
              "validation",
              "calibration",
              "pilot_holdout",
              "final_test"
            ],
            "title": "Split"
          },
          "outcome": {
            "$ref": "#/components/schemas/PrematchOutcomeResponse"
          },
          "margin": {
            "$ref": "#/components/schemas/PrematchMarginResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id",
          "match_date",
          "home_team",
          "away_team",
          "split",
          "outcome",
          "margin"
        ],
        "title": "CombinedPrematchResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InPlayRequest": {
        "properties": {
          "match_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Match Id",
            "examples": [
              267670
            ]
          },
          "minute": {
            "type": "integer",
            "multipleOf": 5.0,
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Minute",
            "description": "Frozen snapshot minute: one of 0, 5, ..., 90.",
            "examples": [
              60
            ]
          },
          "second": {
            "type": "integer",
            "const": 0,
            "title": "Second",
            "description": "The public frozen grid is minute-aligned; only zero is supported.",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id",
          "minute"
        ],
        "title": "InPlayRequest"
      },
...
      "MatchDetail": {
        "properties": {
          "match_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Match Id"
          },
          "competition": {
            "type": "string",
            "const": "La Liga",
            "title": "Competition"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "match_date": {
            "type": "string",
            "format": "date",
            "title": "Match Date"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "split": {
            "type": "string",
            "enum": [
              "train",
              "validation",
              "calibration",
              "pilot_holdout",
              "final_test"
            ],
            "title": "Split"
          },
          "prematch_prediction_available": {
            "type": "boolean",
            "title": "Prematch Prediction Available"
          },
          "inplay_prediction_available": {
            "type": "boolean",
            "title": "Inplay Prediction Available"
          },
          "supported_snapshot_minutes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Supported Snapshot Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id",
          "competition",
          "season",
          "match_date",
          "home_team",
          "away_team",
          "split",
          "prematch_prediction_available",
          "inplay_prediction_available",
          "supported_snapshot_minutes"
        ],
        "title": "MatchDetail"
      },
...
      "MatchesResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "MatchesResponse"
      },
...
      "PrematchMarginResponse": {
        "properties": {
          "match_id": {
            "type": "integer",
            "title": "Match Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "task": {
            "type": "string",
            "const": "prematch_margin",
            "title": "Task"
          },
          "expected_goal_margin": {
            "type": "number",
            "maximum": 5.0,
            "minimum": -5.0,
            "title": "Expected Goal Margin",
            "description": "Expected clipped home-goals minus away-goals margin."
          },
          "derived_outcome_probabilities": {
            "$ref": "#/components/schemas/Probabilities",
            "description": "Probabilities derived by the separately fitted Model 2 margin mapper."
          },
          "model": {
            "type": "string",
            "const": "random_forest",
            "title": "Model"
          },
          "calibrated": {
            "type": "boolean",
            "const": false,
            "title": "Calibrated"
          },
          "probability_mapping": {
            "type": "string",
            "const": "calibration_block_multinomial_logistic",
            "title": "Probability Mapping"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id",
          "home_team",
          "away_team",
          "task",
          "expected_goal_margin",
          "derived_outcome_probabilities",
          "model",
          "calibrated",
          "probability_mapping"
        ],
        "title": "PrematchMarginResponse"
      },
      "PrematchOutcomeResponse": {
        "properties": {
          "match_id": {
            "type": "integer",
            "title": "Match Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "task": {
            "type": "string",
            "const": "prematch_outcome",
            "title": "Task"
          },
          "probabilities": {
            "$ref": "#/components/schemas/Probabilities",
            "description": "Platt-calibrated Model 1 probabilities."
          },
          "raw_probabilities": {
            "$ref": "#/components/schemas/Probabilities"
          },
          "predicted_class": {
            "type": "string",
            "enum": [
              "away_win",
              "draw",
              "home_win"
            ],
            "title": "Predicted Class"
          },
          "model": {
            "type": "string",
            "const": "random_forest",
            "title": "Model"
          },
          "calibrated": {
            "type": "boolean",
            "const": true,
            "title": "Calibrated"
          },
          "calibration": {
            "type": "string",
            "const": "multiclass_platt",
            "title": "Calibration"
          },
          "class_order": {
            "items": {
              "type": "string",
              "enum": [
                "away_win",
                "draw",
                "home_win"
              ]
            },
            "type": "array",
            "title": "Class Order"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id",
          "home_team",
          "away_team",
          "task",
          "probabilities",
          "raw_probabilities",
          "predicted_class",
          "model",
          "calibrated",
          "calibration",
          "class_order"
        ],
        "title": "PrematchOutcomeResponse"
      },
      "PrematchRequest": {
        "properties": {
          "match_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Match Id",
            "description": "StatsBomb match identifier present in the frozen feature store.",
            "examples": [
              267670
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_id"
        ],
        "title": "PrematchRequest"
      },
      "Probabilities": {
        "properties": {
          "home_win": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Home Win"
          },
          "draw": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Draw"
          },
          "away_win": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Away Win"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "home_win",
          "draw",
          "away_win"
        ],
        "title": "Probabilities"
      },
...
    }
  },
...
}